starforge-saas/requirements.txt
# ========================================================
# Connect ATX Elite PaaS – Production-Ready requirements.txt
# Last security audit: July 2025
# All dependencies pinned; CVEs checked with Safety
# Onboarding Tip: Uncomment optional blocks as needed!
# ========================================================

# ───────────── Core Framework ─────────────
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.2

# ────────── Real-time & Async (Pick One) ──────────
Flask-SocketIO==5.3.6
eventlet==0.37.0                # 🔥 SaaS default for async, CVE-patched
# gevent==25.4.2                # Alt: Enterprise async engine (if needed)
# gevent-websocket==0.10.1      # Only if using gevent
redis==5.0.4
Flask-Limiter[redis]==3.12.0

# ─────────── Security & Auth ───────────
Flask-Login==0.6.3              # User auth/session mgmt
Flask-WTF==1.2.1                # Forms, CSRF protection
passlib==1.7.4                  # Secure password hashing

# ───────── Payments & Messaging ─────────
stripe==9.7.0
twilio==9.0.3

# ────────── Storage & File Uploads ──────────
Flask-Uploads==0.2.1            # Image uploads (optional)
# Flask-S3==0.3.4               # S3 integration if needed

# ────────── AI & Monitoring ──────────
openai==1.30.1
sentry-sdk[flask]==2.8.0        # CVE-patched

# ──────── Utility & CLI ────────
python-dotenv==1.0.1
ipython==8.25.0
click==8.1.7                    # CLI commands, Flask integration

# ────── Static Assets & Minification ──────
Flask-Assets==2.1.0             # CSS/JS bundling/minification
cssmin==0.2.0                   # CSS minifier
jsmin==3.0.1                    # JS minifier

# ───── Production WSGI Server ─────
gunicorn==23.0.0                # CVE-patched

# ───── Database Drivers (Pick one) ─────
# psycopg2-binary==2.9.9        # PostgreSQL
# PyMySQL==1.1.1                # MySQL/MariaDB
# sqlite3 (builtin)

# ───── Code Quality & Formatting ─────
black==24.4.2
isort==5.13.2                   # Import sorting

# ────── Testing & Coverage ──────
# pytest==8.2.2
# pytest-cov==5.0.0
# factory-boy==3.3.0

# ────── Email (Optional) ──────
# Flask-Mail==0.9.1

# ────── Image Handling (Optional) ──────
# Pillow==10.3.0

# ────── Background Tasks (Optional) ──────
# celery==5.4.0
# Flask-Celery-Helper==1.3.0

# ────── Admin Interface (Optional) ──────
# Flask-Admin==1.6.1

# ────── Internationalization (Optional) ──────
# Flask-Babel==4.0.0

# ────── API/Serialization (Optional) ──────
# marshmallow==3.21.1
# Flask-RESTX==1.3.0

# ────── Caching (Optional) ──────
# Flask-Caching==2.2.0

# ────── Rate Limiting (already included via Flask-Limiter) ──────

# ────── Monitoring/Profiling (Optional) ──────
# Flask-Profiler==1.10.1

# ────── Sentry, already above ──────

# ────── [Extras] Frontend Build (if using React/Vite etc) ──────
# (Handled by package.json)

# ===== End =====

